moved services and models into new directories and fixed associated imports
