+++ src/app/search-cheese-name/search-cheese-name.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
-import { AllCheeseService } from '../get-all-cheese-btn/getCheesesService';
+import { AllCheeseService } from '../../services/getCheesesService';
 import { Router } from '@angular/router';
-import { CheeseResult} from '../get-all-cheese-btn/cheese.model';
+import { CheeseResult} from '../../models/cheese.model';
 
 @Component({
   selector: 'app-search-cheese-name',
